feat: add new feature implementation
